IKJB42MAIN-5276 arm: msm: fix double mem alloc caused by __va
We have arch specific __phys_to_virt and __virt_to_phys macros
to avoid mapping holes in lowmem. After extraction, the parameter
is placed 3 times and 2 of them are used at runtime. When the
parameter for macro is a memory allocator, it's actually called
twice.
This change updates the macros to use the parameter only once.

Change-Id: I236f66eafea6c9256aed4863d337f9dd1f6dc039
Signed-off-by: Yuanyuan Zhong <a6510c@motorola.com>
Reviewed-on: http://gerrit.pcs.mot.com/521412
Tested-by: Jira Key <jirakey@motorola.com>
Reviewed-by: Check Patch <CHEKPACH@motorola.com>
Reviewed-by: Klocwork kwcheck <klocwork-kwcheck@sourceforge.mot.com>
Reviewed-by: Igor Kovalenko <cik009@motorola.com>
Submit-Approved: Jira Key <jirakey@motorola.com>